Desktop/agent SDK/weather_agent/requirements.txt
openai-agents
chainlit
google-generativeai
python-dotenv
requests